Warning: assert bug
Track recursion in approximations

We now track recursion depth through approximations, which is more
accurate. For example, when unrolling a call to [List.map] whose
argument also contains calls to [List.map], we will now only treat the
actual recursive calls as recursive, the ones coming from the argument
will be considered non-recursive.

To get the right approximations, we mark recursive bindings of functions
using a [Recursive] construct. We let the recursive annotation be
implicit in the renaming map on the set of closures.

We also add a recursion depth field to [Set_of_closures] to handle the
case where we specialise a function with a recursive appoximation.

Note that if a function is non-recursive in a mutually-recursive
set (which of course means the set could be broken up), its calls to
recursive functions aren't truly recursive, so we give them a recursion
depth of zero.