Get recursion depths right in more cases
Specifically: 1) We only want to wrap a recursive call in a Recursive
declaration if it's a recursive function body calling itself or a recursive
sibling. 2) When replacing a Move_within_set_of_closures, we may need to wrap it
in a *negative* Recursive declaration if the set of closures has a recursion
depth, so that the function call ends up at 0 recursion depth again.