Changes in anticipation of the 4.12 set of closure repr changes (#269)
* Add option to print the computed offsets

* Check that repr of closure match the new one since 4.12

Starting from ocaml 4.12, closure representaiton changed slightly (see
PR#9619 on the upstream github repo). This commit adds checks in
un_cps_closure to ensure that the computed offsets are conform to the
new representation of closures, mainly that all closure slots precede
all env vars slots, so that the start of environment can be defined.

* Update layout type to contain start of env

In anticipation of the rebase on 4.12, compute the start of env for a
set of closures and update the layout type to contain that information.
This will greatly help when rebasing.

* Printer updates

* Add code to enforce set of closure invariant

This code will ensure that env var slots are always assigned slots that
are after closure slots (and hence will not take advantage of "holes"
between closure slots). Note that while this should prevent set of
closures defined in the same compilation unit from creating conflicts
when computing offsets, constraints from other compilation units can
still make some situation unsatisfiable.

* Printing update

* WIP: add example testing offset computation with sharing

* Update middle_end/flambda/to_cmm/un_cps_closure.ml

Co-authored-by: Vincent Laviron <vincent.laviron@gmail.com>

* Address reviews

* Update middle_end/flambda/to_cmm/un_cps_closure.ml

Co-authored-by: Vincent Laviron <vincent.laviron@gmail.com>

* Update middle_end/flambda/to_cmm/un_cps_closure.ml

Co-authored-by: Vincent Laviron <vincent.laviron@gmail.com>

Co-authored-by: Vincent Laviron <vincent.laviron@gmail.com>