Revert [Code_metrics.expr_size].
This function was supposed to be called frequently in a previous
iteration of this feature. For performances it was refactored to be
tail recursived but in the end this function is only called twice in
the codebase:
- once to test that the size present in uacc is the same as the one
given by the explicit computation with [expr_size]
- once in the inliner and this call will be soon removed.

As the refactor was buggy it's better to just revert it.