Depth variables (#427)
This adds a new sort of variable that stands for some unknown recursion
depth. Roughly speaking, it's a variable that holds a `Rec_info`. This
patch doesn't include anything that actually uses depth variables, just
the plumbing to keep track of them. The next patch will add a depth
variable as a parameter to each function.