Pass coercion to Simple.pattern_match argument (#457)
This patch doesn't fix any broken behavior, but it will make it easier
to find. I've looked at each call to `Simple.pattern_match`, and
whenever I couldn't be sure we're not dropping a coercion, I put in a
CR (for easy grepping, always either "Coercion dropped!" or "Coercion
dropped?").