Allow building FLambda parser in Dune (#495)
Unfortunately, I couldn't find a nice way to regenerate
`flambda_parser.ml`, etc. conditionally, which is important because it
requires not just Menhir but precisely the correct version of
Menhir (the same one the library was built with). So one has to change
the mode from `fallback` to `promote` manually. But after that, you get
all the benefits of `dune build -w` when editing the parser or lexer.